add MasterAdvancedChannel trait instance
